requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
langchain>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.7.0
yfinance>=0.2.36
requests>=2.31.0
gunicorn==21.2.0
